+++ src/pydvl/valuation/scorers/__init__.py
@@ -13,6 +13,6 @@ for a certain quality of approximation.
 
 from .base import *
 from .classwise import *
+from .skorch import *
 from .supervised import *
-from .torchscorer import *
 from .utils import *
